refactoring traveling salesperson code
